fix: Delete correct record
The function that deletes records did not include the text value field
when finding the record to delete. Sometimes multiple records were found
and then this function was not able to delete them.

Make sure the right record is found by also including the record type
and text.

This fixes issue #1.